feat!: Replaced EncryptedConnection#listener with EncryptedClient#listener
feat: Added clientside-event for EncryptedClient.java#disconnect

chore: cleaned up some code
