+++ src/main/java/de/pterocloud/encryptedconnection/listener/ServerListener.java
@@ -6,20 +6,7 @@ import de.pterocloud.encryptedconnection.Packet;
 
 import java.net.Socket;
 
-/**
- * The Listener for an entire EncryptedServer
- * Trigger: PacketReceived, PreConnect (not encrypted), PostConnect (encrypted)
- */
 public interface ServerListener {
-
-    /**
-     * Triggered when a packet is received
-     *
-     * @param connection the connection
-     * @param packet     the packet
-     */
-    default void onPacketReceived(EncryptedConnection connection, Packet<?> packet) {
-    }
 
     /**
      * Triggered before the connection is encrypted
@@ -38,6 +25,26 @@ public interface ServerListener {
      * @param connection the connection
      */
     default void onPostConnect(EncryptedClient client, EncryptedConnection connection) {
+
+    }
+
+    /**
+     * Triggered when the connection is disconnected
+     *
+     * @param connection the connection
+     */
+    default void onDisconnect(EncryptedClient client, EncryptedConnection connection) {
+
+    }
+
+    /**
+     * Triggered when a packet is received
+     *
+     * @param connection the connection
+     * @param packet     the packet
+     */
+    default void onPacketReceived(EncryptedConnection connection, Packet<?> packet) {
+
     }
 
 }
